requirements.txt
typing-extensions==4.7.1
tensorflow==2.8.0
optax==0.1.7
dm-env==1.6
dm-launchpad==0.5.2
jax==0.4.14
jaxlib==0.4.14
dm-haiku==0.0.10
numpy==1.24.3
rlax==0.1.6
absl-py==1.4.0
dm-reverb==0.7.2
dm-sonnet==2.0.1
rlds==0.1.8
flax==0.7.2
tensorflow_probability==0.15.0
dm-tree==0.1.8
scipy==1.10.1
protobuf==3.20.3
scikit-image==0.21.0
pymunk==6.5.1
pybind11==3.0.0
